requirements.txt
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0

# GNN Enhancement Dependencies
torch>=1.9.0
dgl>=0.9.0
networkx>=2.6.0
torch-geometric>=2.0.0

# Core Infrastructure Dependencies
pyyaml>=6.0.0
cryptography>=3.4.8
psutil>=5.9.0

# Testing Framework
pytest>=7.1.0
pytest-cov>=3.0.0

# Code Quality Tools
black>=22.6.0
flake8>=4.0.0
mypy>=0.971
pylint>=2.14.0
bandit>=1.7.4
safety>=2.1.0

# Natural Language Processing
nltk>=3.7.0
regex>=2022.6.2

# Configuration and Logging
python-dotenv>=0.20.0
loguru>=0.6.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Optional Advanced Features (uncomment as needed)
# fastapi>=0.78.0
# uvicorn>=0.18.0
# redis>=4.3.0
# sqlalchemy>=1.4.0
